moved authOptions from api to lib because Vercel error
