+++ src/lib/authOptions.ts
@@ -1,0 +1,40 @@
+// lib/authOptions.ts
+import { type NextAuthOptions } from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
+import bcrypt from "bcryptjs";
+import { sql } from "./dbCon";
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "text" },
+        password: { label: "Password", type: "password" },
+      },
+      async authorize(credentials) {
+        const user = await sql`
+          SELECT * FROM users WHERE email = ${credentials?.email}
+        `;
+        if (user.length === 0) return null;
+
+        const valid = await bcrypt.compare(credentials!.password, user[0].password_hash);
+        if (!valid) return null;
+
+        return { id: user[0].id.toString(), email: user[0].email };
+      },
+    }),
+  ],
+  session: { strategy: "jwt" },
+  pages: { signIn: "/login" },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) token.id = user.id;
+      return token;
+    },
+    async session({ session, token }) {
+      if (session.user) session.user.id = token.id as string;
+      return session;
+    },
+  },
+};
